+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- [**BC**] Use net8.0
+- Fix package metadata
+- Serialize numbers in json value to `int64`
 
 ## 5.0.0 - 2023-09-09
 - [**BC**] Use `Alma` namespace